requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
autopep8==1.4.4
cffi==1.12.3
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
dj-static==0.0.6
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.4.1
parsel==1.5.2
pep8==1.7.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pytz==2019.2
queuelib==1.5.0
Scrapy==1.7.3
scrapy-djangoitem==1.1.1
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
Twisted==19.7.0
w3lib==1.21.0
zope.interface==4.6.0
